requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==2.1.11
Django==4.0.10
django-allauth==0.48.0
django-cache-url==3.4.5
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-spectacular==0.21.2
environs==9.5.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
marshmallow==3.21.2
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.1
requests-oauthlib==2.0.0
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==5.3.0
